Get font names corresponding to four style names for a font family

systemfonts::system_fonts() などの結果から4書体のフォントを取 得する. ただしうまくいくとは限らない.

get_styles(data)

Arguments

data

data.frame alike, output of font list from systemfonts:: system_fonts() or its subset, which includes six columns: name, family, style, weight, width, and italic/ data.frame 相当のオブジェクトで, systemfonts::system_fonts() のから特定のファミリだけ取り出したもので, name, family, style, weight, width, italic 列を持つもの.

Value

character vector of length 4: It's elements are names of plain, bold, italic, and bold italic font./長さ4の文字列ベクトル. 要素は順に標準, ボールド, イタリック, ボールドイタリック体のpostscript名.

Details

The original function was transferred from https://github.com/Gedevan-Aleksizde/fontregisterer (Version 0.3) accessed on 2023/06/02 (Fri.).