To display two font families in a single character string on ggplot title and labels

This function has been made to cope with VERY VERY specific demand in Japanese domestic Stock assessment.

mixfonts(strs, asciifont = "Arial", mbytefont = "MS Gothic")

Arguments

strs
A vector of character to be converted

asciifont
ASCII font family, Default: 'Arial'

mbytefont
Multi byte font family, Default: 'MS Gothic'

Value

An HTML span tag string.

Details

This function is supposed to be used with element_markdown() in ggtext package

See also

Examples

if (FALSE) {
if(interactive()){
 library(ggplot2)
 library(ggtext)

 # EXAMPLE1
 mixfonts(c("シラス込み0歳魚", "\nシラスを除く\nカタクチイワシ0歳魚"))

 # EXAMPLE2
 ggplot(women) +
     geom_point(aes(x = height, y = weight)) +
     labs(x = mixfonts("身長\n(inch)", asciifont = "Times New Roman")) +
     theme_linedraw(base_family = "MS Gothic") +
     theme(axis.title.x.bottom = element_markdown())
 }
}